refted files name(camelCase -> snake_case), fixed bugs with git branch in command prompt, refted command Histories logic. added Beta version now orbix not afraid of panics example: ORB C:\Users\Admin\MyCMD\goCMD\scripts>PANIC "test panic"
Panic recovered: test panic
Recovering from panic

Successfully recovered from the panic: test panic
ORB C:\Users\Admin\MyCMD\goCMD\scripts>

added more logic to server recover and Orbix recover

+++ cmd/commands/commandsWithSignaiture/Remove/remove.go
@@ -10,6 +10,7 @@ import (
 func File(command string, commandArgs []string) (string, error) {
 	yellow := color.New(color.FgYellow).SprintFunc()
 	green := color.New(color.FgGreen).SprintFunc()
+	red := color.New(color.FgRed).SprintFunc()
 
 	if len(commandArgs) < 1 {
 		fmt.Println(yellow(fmt.Sprintf("Usage: %s <file>", command)))
@@ -18,17 +19,22 @@ func File(command string, commandArgs []string) (string, error) {
 
 	name := commandArgs[0]
 
+	if name == "running.txt" {
+		fmt.Println(red("You can not delete file \"running.txt\""))
+		return "", nil
+	}
+
 	if err := os.Remove(name); err != nil {
 		if strings.Contains(err.Error(), "being used by another process") {
-			fmt.Printf("Error: Cannot remove '%s' because it is being used by another process.\n", name)
+			fmt.Println(red(fmt.Sprintf("Error: Cannot remove '%s' because it is being used by another process.\n", name)))
 		} else {
-			fmt.Printf("Error removing file: %v\n", err)
+			fmt.Println(red(fmt.Sprintf("Error removing file: %v\n", err)))
 		}
 		return name, err
 	}
 
 	if name != "" {
-		fmt.Println(green(fmt.Sprintf("File '%s' successfully removed.\n", name)))
+		fmt.Println(green(fmt.Sprintf("File '%s' successfully deleted.\n", name)))
 	}
 
 	return name, nil
